Outlook/requirements.txt
azure-identity==1.15.0
msgraph-core==1.0.0
pandas==2.3.1
openpyxl==3.1.2
python-dotenv==1.0.1
msal==1.26.0
tqdm==4.66.2
requests==2.31.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
types-requests==2.31.0.20240125
types-tqdm==4.66.0.5 